Add tests for `(schemaMap).AttributeByPath()` and `(schemaMap).BlockByPath()`
